extract jax dependencies
